requirements.txt
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
datasets>=2.14.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0